pyproject.toml
[tool.poetry]
name = "Poetry config"
version = "0.1.0"
description = ""
authors = ["Your Team <your@gmail.com>"]

[tool.poetry.dependencies]
python = "3.12.0"
fastapi = "^0.112.2"
uvicorn = "0.16.0"
gunicorn = "^22.0.0"
environs = "^9.3.5"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
loguru = "^0.5.3"
python-slugify = "^5.0.2"
furl = "^2.1.3"
slugify = "^0.0.1"
itsdangerous = "^2.2.0"
httpx = "^0.22.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
uuid = "^1.30"
email-validator = "^2.2.0"
python-dateutil = "2.9.0.post0"
aio-pika = "^9.4.3"

[tool.poetry.dev-dependencies]
pytest = "^6.1.0"
pytest-asyncio = "^0.18.3"
pytest-cov = "^3.0.0"
flake8 = "^7.1.1"
mypy = "^1.13.0"
Sphinx = "^4.3.2"
black = "^22.12.0"
sphinx-rtd-theme = "^1.0.0"
sphinx-autoapi = "^1.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "-p no:warnings --asyncio-mode=strict"
testpaths = ["tests"]
asyncio_mode = "auto"


[tool.black]
line-length = 115
target-version = ['py311']

[tool.mypy]
python_version=3.12
disallow_untyped_defs=1
ignore_missing_imports=1
